% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_charts.R
\name{create_geom_col_chart}
\alias{create_geom_col_chart}
\title{Creates a bar graph using geom_col().}
\usage{
create_geom_col_chart(attribute, data, entity)
}
\arguments{
\item{attribute}{The name of the categorical variable in data by which the
bar graph is to categorize the data.}

\item{data}{A data.frame containing the data from an R dataset.}

\item{entity}{The name of what the data in the dataset represents (for
example, "Students" for HairEyeColor).}
}
\value{
A ggplot object for the bar graph.
}
\description{
Creates a bar graph categorizing the data in a data.frame based on an R
dataset according to values of the categorical variable named attribute,
where the height of each bar is a value in the data.frame.
}
