% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tab.R
\name{create_tab}
\alias{create_tab}
\title{Creates a new tab for a navlistPanel.}
\usage{
create_tab(tab_name, select_name, attributes, plot_name)
}
\arguments{
\item{tab_name}{The name of the new tab.}

\item{select_name}{The name of the selectInput.}

\item{attributes}{A vector of attribute strings for the selectInput.}

\item{plot_name}{The name of the bar graph.}
}
\value{
A new tabPanel.
}
\description{
Creates a new tab displaying a bar graph for a dataset. The tab has a
selectInput for selecting one of the attributes in the attributes vector
parameter. The selected attribute is the categorical variable according to
which data from the dataset is categorized to plot a bar graph.
}
